laporan chart unggah berkas selesai

+++ database/migrations/2022_07_10_011328_create_mhs_temp_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('mhs_temp', function (Blueprint $table) {
-            $table->bigInteger('no_online');
+            $table->integer('no_online');
             $table->string('no_test');
             $table->string('nim');
             $table->string('jalur_masuk');
